POA: Remove 2 redundant type casts
Removed unnecessary type casts where compiler can infer the type:
- TenantController.cs: removed decimal cast on StorageUsedGB multiplication (line 170)
- RevocationRegistryService.cs: removed OcspResponseStatus cast (line 377)

The multiplication already produces decimal type, and the status parameter
is already of OcspResponseStatus type, making the explicit casts redundant.

+++ src/NumbatWallet.Web.Api/Controllers/TenantController.cs
@@ -167,7 +167,7 @@ public class TenantController : ControllerBase
                 WalletCount = stats.TotalWallets,
                 CredentialCount = stats.TotalCredentials,
                 ActiveSessions = 0, // Not tracked in the DTO
-                StorageUsedMB = (decimal)(stats.StorageUsedGB * 1024), // Convert GB to MB
+                StorageUsedMB = stats.StorageUsedGB * 1024, // Convert GB to MB
                 LastActivityDate = stats.PeriodEnd
             };
